This describes changes from calcos version 2.15.4 to 2.17.
Relevant Trac tickets and PRs:
Trac   PR
----  -----
 575  65353  Calcos should have an option to search for the target spectrum.
 645  67204  Add a TIMELINE extension to calcos corrtag files.
 719  68545  Calcos needs an option to use the default location for the
             target spectrum.
 720  68550  Calcos should accept data taken with different cenwaves.
 761  69143  Calcos should flag gain-sagged regions in FUV data.
 763  69206  Calcos should check for no rows in G140L lampflash table.
 764  69207  Calcos needs a utility function for segment-specific keywords.
 786  69401  Calcos should extrapolate the TDS correction.
 789  69437  Calcos should expand wildcards for the input file names.
 791  69679  Calcos should use the shift file if it's relevant.
 792  69694  Remove the test in calcos on allowed values of OPT_ELEM.
 793  69730  In calcos, modify some pyfits.open calls.
 795  69737  Change the way calcos fixes APERTURE = "RelMvReq".

An association table may now include exposures with different values
of central wavelength.  This should not happen in the OPUS pipeline, but
a user can create such an association table.  The grating must still be
the same for every input file in an association.

The TIMELINE table in the corrtag file has a new column, SUN_ZD, the angle
between HST and the Sun, as seen from the center of the earth.  This is
approximately equal to the zenith distance of the Sun as seen from HST,
except that the parallax of the Sun is not included.

A new reference table GSAGTAB may be specified.  This "gain sag" table is
a bad pixel table that flags regions of the FUV detector that have very low
gain.

Previously, calcos would accept multiple input file names, but there was
a limit due to the line length.  The input file or files may now be
specified (in quotes) using wildcard characters.

The syntax for the command-line option --find has been changed; a value
should also be specified.  This option has to do with finding the target
spectrum in the cross-dispersion direction.  The value can be "yes", "no",
or a numerical value, a cutoff.  ("true" and "false" are aliases for "yes"
and "no" respectively.)  If a cutoff value is given, this means that the
spectrum will be extracted at the found location if the error estimate
for the location is less than or equal to the cutoff value.

When finding the cross-dispersion location of a spectrum, the quadratic
fit was being done to too few pixels (at least for FUV), resulting in a
poor value for the standard deviation of the fit.  The number of pixels
will now be based on the FWHM, or a minimum of five pixels.

The command-line option --shift <file> did not work for the case that the
raw file was not tagflash and there was no auto/GO wavecal.  The code
expected tagflash data, and the shift file would specify which lamp flash
was to be overridden.  This has been modified so that if a shift file
was specified for non-tagflash data, the given shift will be applied.

The extrapolation of the time-dependent sensitivity factors has been
modified.  Previously, if the time of observation were before the first
or after the last time in the TDS table, the factor would be extrapolated
with zero slope.  Now the extrapolation will be done using the slope at
the first or last time, as appropriate.  Note, however, that the TDS table
should cover all observations without extrapolation.

Changes were made to some pyfits.open statements to allow using memory
mapping.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@13992 fe389314-cf27-0410-b35b-8c050e845b92