Relevant Trac tickets and PRs: Trac   PR ----  -----  720  68550  Calcos should accept data taken with different cenwaves.  761  69143  Calcos should flag gain-sagged regions in FUV data.  805  69937  Account for the width of the OI 1304 line in calcos.  829  70033  Calcos should check whether a keyword exists before deleting it.  830  70036  Calcos should read the HVTAB and populate keywords.
The changes to disable use of the gain sag table (gsagtab) are:
    in calcos.py, remove gsagtab from missingRefFiles (two places)
    in getinfo.py, explicitly set gsagtab in reffiles to "N/A"
Look for the string "xxx temp disable".

Changes to calcos.py

The option to combine (average) data taken with different cenwaves in
the same association has been disabled.  This was done by modifying
compareConfig to put back the test on cenwave, and by changing
missingRefFiles to remove gsagtab from the `temp` dictionary and removing
the test for existence of the gsagtab just after checking for bpixtab.
Trac #720, PR 68550

Changes to cosutil.py:

Function findGSagExtn was revised to use keyword HVLEVELA or HVLEVELB.
Trac #830, PR 70036

Changes to getinfo.py:

Function getRefFileNames was modified to set reffiles["gsagtab"] to "N/A",
regardless of whether there's a gsagtab keyword in the header or not.
The two lines to do this are flagged with "# xxx temp disable".
Trac #761, PR 69143

Keyword HVLEVELA or HVLEVELB is gotten instead of DETHVL[AB].  The name of
the reference file for FUV high voltage (keyword HVTAB) is gotten.
Trac #830, PR 70036

Changes to spwcs.py:

Method deleteKeywords was modified to check whether a keyword is actually
present in the header before deleting it.
Trac #829, PR 70033

Changes to timeline.py:

Function createTimeline was modified to print the range of pixel numbers
in X and in Y for airglow regions ly_alpha and oi_1304 (if verbose).
Trac #805, PR 69937

Changes to timetag:

Function updateHVKeywords was added, and this is called close to the
beginning of timetagBasicCalibration, just after cosutil.overrideKeywords.
Trac #830, PR 70036


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@14499 fe389314-cf27-0410-b35b-8c050e845b92