Relevant Trac tickets and PRs: Trac   PR ----  -----  616  66158  Calcos should be modified to use Sphinx for documentation.  960  72794  Calcos should do the Doppler correction after csum.
Overview:

The Doppler correction for TIME-TAG data is now done after the call to
write the csum file.  The Doppler correction uses the dispersion and
1-D extraction reference tables (DISPTAB and XTRACTAB), and there may
be situations where these tables do not contain rows for some mode,
but it is desired to do some calibration (e.g. GEOCORR) and write a csum
file before stopping with an exception.  It was possible to move Doppler
correction because the csum file does not depend on it.

There should be no change in the output of version 2.19.5 compared with
2.19.4 except that the primary header keyword DOPPCORR will now be
'PERFORM' instead of 'COMPLETE' in csum files.

Details:

The following files were modified to remove backquotes:
accum.py, calcos.py, concurrent.py, cosutil.py, extract.py, findshift1.py,
fpavg.py, timeline.py, timetag.py, wavecal.py, x1d.py, xd_search.py.
This was prompted by an error building the documentation.
Trac #616, PR 66158

Changes to timetag.py:

The call to doDoppcorr was moved from just before the call to doDeadcorr
to a point after the call to writeCsum.
Trac #960, PR 72794


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@20677 fe389314-cf27-0410-b35b-8c050e845b92