The following files were modified to remove backquotes: accum.py, calcos.py, concurrent.py, cosutil.py, extract.py, findshift1.py, fpavg.py, timeline.py, timetag.py, wavecal.py, x1d.py, xd_search.py. This was prompted by an error building the documentation. Trac #616, PR 66158
Changes to calcos.py:

Within the loop over input files, a trap was added for either a bad
aperture name or a missing table row exception.  If an error was found,
an error message will be printed, but the loop will then continue to try
to process the next file.
Trac #960, PR 72794

The option to combine (average) data taken with different cenwaves in
the same association has been disabled.  This was done by modifying
compareConfig to put back the test on cenwave.
Trac #720, PR 68550

Changes to calcosparam.py:

A new MissingRowError exception is defined.  This allows trapping this
condition (see calcos.py) in order to continue processing all the input files.
Trac #960, PR 72794

Changes to concurrent.py:

ConcurrentWavecal.avgShift() was modified to normalize by the full range
of time covered by the exposure, i.e. the last value in the TIME column
minus the first value in the TIME column, because that's the range of time
used when incrementing the sums for computing averages.
Trac #963, PR 72824

Changes to timetag.py:

The calls to doPhacorr and doDoppcorr were moved to a point after the
call to writeCsum, and the call to doBurstcorr was moved to just after
wavelength calibration.
Trac #960, PR 72794

In deadtimeCorrectionAccum, if the exposure time is zero, the returned
tuple of dummy values now has the correct length of three (instead of two).
In convolveFlat, npts is set to at least 1, and npts is now used both for
the length of the t array as well as the range of the loop over i.  This
avoids dividing by zero and an index-out-of-bounds error if the exposure
time is zero.
Trac #960, PR 72794


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@20687 fe389314-cf27-0410-b35b-8c050e845b92