Relevant Trac tickets and PRs: Trac   PR ----  -----  718  68538  Update deprecated Python syntax in calcos.  735  68712  Change calcos import statements for boxcar.  961  72796  Calcos gives the wrong EXPTIME in x1dsum for NUV.  973  73085  Calcos should check wavecorr before updating WAVECALS.  981  73424  Calcos should use a SEARCH_OFFSET column in the WCPTAB.  982  73439  Calcos should allow aperture = "N/A".  984  73571  Calcos x1dsum files have DQ = 0 for some bad pixels.  988  73686  Calcos should import pyfits from astropy.  995  74005  Function in calcos gets DETECTOR from wrong header. 1001  74177  Calcos should check that EXPEND - EXPSTART is not zero. 1019  75005  Have CalCOS check the exposure time (EXPTIME) before the CSUMs              are created
The following files were modified regarding importing pyfits (specifically,
"import astropy.io.fits as fits" instead of "import pyfits"):
accum.py, average.py, calcos.py, concurrent.py, cosutil.py, extract.py,
fpavg.py, getinfo.py, orbit.py, splittag.py, spwcs.py, timeline.py,
timetag.py, wavecal.py, x1d.py
Trac #988, PR 73686

The following files were modified to replace deprecated syntax in the
astropy.io.fits package:
accum.py, average.py, calcos.py, concurrent.py, cosutil.py, extract.py,
fpavg.py, phot.py, splittag.py, spwcs.py, timeline.py, timetag.py,
wavecal.py, x1d.py
These three forms were replaced:
    header.update(keyword, value)
    header.update(keyword, value, comment=comment)
    header.update(keyword, value, after=after_keyword)
Trac #718, PR 68538

Changes to accum.py:

The check on whether the aperture name is in the list of recognized values
has been extended to include "N/A", which is valid for dark exposures.
Trac #982, PR 73439

Changes to concurrent.py:

Function processConcurrentWavecal was modified to add wavecorr to the
tuple that is returned.  This can be used to check whether the wavecal
shifts were determined from the data (even if copied from segment A to B),
as opposed to being assigned a default value.
Trac #973, PR 73085

Function findShifts was modified to get the value from column SEARCH_OFFSET
in the wavecal parameters table (if that column is present) and add it to
initial_offset.
Trac #981, PR 73424

Changes to cosutil.py:

Function computeLifeAdjOffset assigns a value to info["life_adj_offset"],
based on keywords life_adj and aperypos (and other parameters).  This
should be done if life_adj is -1, but the test allowed any zero or
negative value.  This test was changed to specifically require life_adj
to be -1 before setting info["life_adj_offset"] to a non-zero value.
Trac #982, PR 73439

Changes to findshift1.py:

__init__() was modified to allow initial_offset to be non-zero even if
columns FPOFFSET and FP_PIXEL_SHIFT are present in the lamptab.  The data
type of initial_offset had been specified in the doc string to be int, but
the value may be float, so the doc string was modified.
Fixed a typo in the doc string for repairNUV:  "1.5 pixels" should have been
"1.2 pixels".
Trac #981, PR 73424

Changes to fpavg.py:

For NUV data, getInputInfo was assigning a value to the EXPTIME keyword
that was three times too large.  This has been fixed.  The values in the
EXPTIME column are not affected.
Trac #961, PR 72796

Function convolveFlat was modified to avoid dividing by zero for the case
that EXPSTART and EXPEND are the same.  A variable is now assigned the
value of EXPEND - EXPSTART, rounded to an integer; if this value is less
than or equal to zero, the input flat will be returned unmodified.
Trac #1001, PR 74177

Function getDopplerParam was trying to get DETECTOR from the extension
header, but the keyword is in the primary header.  This keyword is now
gotten by Spectrum.__init__ and passed to getDopplerParam.
Trac #995, PR 74005

Function fpInitData was modified to initialize the DQ array with
DQ_PIXEL_OUT_OF_BOUNDS instead of zero.
(In calcos 2.20.0, this was handled by changing normalizeSums to set DQ to
DQ_PIXEL_OUT_OF_BOUNDS if DQ_WGT was not positive, but this resulted in
some pixels not near an edge being flagged with this value, 128).
Trac #984, PR 73571

Changes to timetag.py:

A call to recomputeExptime was added before calling writeCsum, to ensure
that the exposure time keyword would be correct (to the extent that we
know the value) in the case that calcos failed (e.g. due to a bad value
of the APERTURE keyword) not long after writing the csum file.
Trac #1019, PR 75005

Function timetagBasicCalibration checks the wavecorr variable that is
now returned by concurrent.processConcurrentWavecal(), and the WAVECALS
keyword is only updated for tagflash data if wavecorr is "COMPLETE".
For non-tagflash data, function updateFromWavecal checks wavecorr and only
update keyword WAVECALS if wavecorr is "COMPLETE".
Trac #973, PR 73085

The check on whether the aperture name is in the list of recognized values
has been extended to include "N/A" for dark exposures.
Trac #982, PR 73439

Changes to wavecal.py:

Function findWavecalShift was modified to get the value from column
SEARCH_OFFSET in the wcptab (if that column is present) and add it to
initial_offset.
Trac #981, PR 73424

Changes to wavecal.py and xd_search.py:

The import statement for boxcar was in a try/except block; this was
changed to an unconditional import from stsci.convolve.
Trac #735, PR 68712


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/calcos/trunk@26502 fe389314-cf27-0410-b35b-8c050e845b92